feat : implement filters and errors and try testing using jest

+++ okr-server/src/objective/objective.service.ts
@@ -1,28 +1,65 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from '../prisma.service';
 import { Objective } from '../../generated/prisma/client';
-
-export interface ObjectiveDtoType {
-  objective: string;
-}
+import { ObjectiveDtoType } from './dto/objectiveDtoType';
+import { ObjectiveNotFoundError } from './error/objectiveNotFoundError';
+import { DuplicateObjectiveError } from './error/duplicateObjectiveError';
 
 @Injectable()
 export class ObjectiveService {
-  constructor(private prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService) {}
 
   fetchAll(): Promise<Objective[]> {
-    return this.prisma.objective.findMany();
+    return this.prisma.objective.findMany({
+      include: { keyResults: true },
+    });
   }
 
-  create(data: ObjectiveDtoType): Promise<Objective> {
+  async create(data: ObjectiveDtoType): Promise<Objective> {
+    const objective = await this.prisma.objective.findFirst({
+      where: { objective: data.objective },
+    });
+    if (objective) {
+      throw new DuplicateObjectiveError(objective.id, data);
+    }
     return this.prisma.objective.create({ data });
   }
 
-  delete(id: string): Promise<Objective> {
+  async delete(id: string): Promise<Objective> {
+    const objective = await this.prisma.objective.findUnique({
+      where: { id },
+    });
+
+    if (!objective) {
+      throw new ObjectiveNotFoundError(id);
+    }
     return this.prisma.objective.delete({ where: { id } });
   }
 
-  update(id: string, data: ObjectiveDtoType): Promise<Objective> {
+  async update(id: string, data: ObjectiveDtoType): Promise<Objective> {
+    const objective = await this.prisma.objective.findUnique({
+      where: { id },
+    });
+
+    if (!objective) {
+      throw new ObjectiveNotFoundError(id);
+    } else if (objective.objective === data.objective) {
+      throw new DuplicateObjectiveError(id, data);
+    }
+
     return this.prisma.objective.update({ where: { id }, data });
+  }
+
+  async getById(objectiveId: string) {
+    const objective = await this.prisma.objective.findUnique({
+      where: { id: objectiveId },
+      include: { keyResults: true },
+    });
+
+    if (!objective) {
+      throw new ObjectiveNotFoundError(objectiveId);
+    }
+
+    return objective;
   }
 }
